ASR/host/v2/requirements.txt
flask
flask-cors
werkzeug
torch
nemo_toolkit[asr]
omegaconf
librosa
soundfile
numpy
psutil